+++ src/components/calendar.ts
@@ -1,0 +1,2 @@
+export * from './calendar/calendar.component.js';
+export { default } from './calendar/calendar.component.js';